+++ UScript.Runtime/metadata.h
@@ -16,6 +16,6 @@ USCRIPT_ERR create_mdctx(char *, UScriptMetadataContext **);
 USCRIPT_ERR parse_func_tbl(char *, FunctionMetadataTable **, UScriptMetadataContext *);
 
 //! Resolves a function token and returns it's metadata row.
-USCRIPT_ERR resolve_func_token(FunctionMetadataRow **row, UScriptMetadataContext *ctx, int64_t token);
+USCRIPT_ERR resolve_func_token(FunctionMetadataRow **row, UScriptMetadataContext *ctx, int32_t token);
 
 #endif
